requirements.txt
configparser
requests
lxml
python-dateutil
